requirements.txt
moviepy==1.0.3
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
SoccerNet==0.1.60
timm==1.0.11
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
wandb==0.18.5